requirements.txt
alembic==1.13.1
async-generator==1.10
attrs==23.1.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
exceptiongroup==1.1.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
numpy==1.24.3
outcome==1.2.0
pandas==2.0.2
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==3.0.2
wsproto==1.2.0
